Create DistributionBasePlugin
This is essentially the distribution plugin, except it does not create a main distribution by default.
Most logic from distribution was moved to distribution-base
The distribution plugin applies distribution-base

DistributionBasePlugin is introduced as non-incubating.
Release notes entry has been added.
Userguide has been updated to mention the base plugin.

The existing integ tests have been split between those that operate on the main distribution and those that operate on custom distributions.
