+++ testing/architecture-test/src/changes/accepted-changes/accepted-public-api-changes.json
@@ -1,6 +1,20 @@
 {
     "acceptedApiChanges": [
         {
+            "type": "org.gradle.api.distribution.plugins.DistributionBasePlugin",
+            "member": "Class org.gradle.api.distribution.plugins.DistributionBasePlugin",
+            "acceptation": "No need to incubate",
+            "changes": [
+                "Interface has been added"
+            ]
+        },
+        {
+            "type": "org.gradle.api.distribution.plugins.DistributionPlugin",
+            "member": "Constructor org.gradle.api.distribution.plugins.DistributionPlugin()",
+            "acceptation": "No need to incubate",
+            "changes": []
+        },
+        {
             "type": "org.gradle.api.plugins.SoftwareReportingTasksPlugin",
             "member": "Class org.gradle.api.plugins.SoftwareReportingTasksPlugin",
             "acceptation": "Splitting diagnostics into base- and software-",
@@ -36,6 +50,14 @@
             "acceptation": "Protected methods intended to be overriden by Gradle, not API",
             "changes": [
                 "Method is now abstract"
+            ]
+        },
+        {
+            "type": "org.gradle.kotlin.dsl.BuiltinPluginIdExtensionsKt",
+            "member": "Method org.gradle.kotlin.dsl.BuiltinPluginIdExtensionsKt.getDistribution-base(org.gradle.plugin.use.PluginDependenciesSpec)",
+            "acceptation": "Generated extension",
+            "changes": [
+                "Method added to public class"
             ]
         },
         {
